Update index.md
I modified the DSOMM landing page to accommodate a broader audience than those who were there for the presentation this was intended for. I also modified the language a little to be hopefully more descriptive of DSOMM. Lastly, I added a few additional links for English speakers to find more information on DSOMM and added myself as a "Project Leader" so that I can help support any AMER based contributions or requests. Thanks!